fix(ui): clear provider credential drafts when switching agents
