+++ ui/src/e2e/model-provider-usage-outcomes.e2e.test.ts
@@ -6,7 +6,7 @@ import { installMockGateway } from "../test-helpers/control-ui-e2e.ts";
 import { createControlUiE2eSuite } from "./control-ui-e2e-suite.test-support.ts";
 
 const suite = createControlUiE2eSuite({
-  name: "Control UI Model Provider usage outcomes mocked Gateway E2E",
+  name: "Control UI Model Provider operator outcomes mocked Gateway E2E",
   startServerBeforeBrowser: true,
   unavailableMessage: (executablePath) =>
     `Playwright Chromium is not available at ${executablePath}`,
@@ -109,6 +109,111 @@ suite.define(() => {
           await card.screenshot({
             animations: "disabled",
             path: path.join(artifactDir, "provider-usage-provider-error.png"),
+          });
+        }
+      },
+    );
+  });
+
+  it("clears unsaved provider credentials when switching agents without saving them", async () => {
+    await suite.withPage(
+      {
+        locale: "en-US",
+        serviceWorkers: "block",
+        viewport: { height: 1_000, width: 1_280 },
+      },
+      async ({ page }) => {
+        const config = {
+          agents: { defaults: { model: "openai/gpt-5.5" } },
+          models: { providers: { openai: { apiKey: "[redacted]" } } },
+        };
+        const gateway = await installMockGateway(page, {
+          defaultAgentId: "main",
+          featureMethods: ["chat.metadata", "chat.startup", "config.patch"],
+          methodResponses: {
+            ...providerUsageResponses({ updatedAt: now, providers: [] }),
+            "agents.list": {
+              agents: [
+                { id: "main", name: "Main" },
+                { id: "writer", name: "Writer" },
+              ],
+              defaultId: "main",
+              mainKey: "main",
+              scope: "agent",
+            },
+            "config.get": {
+              config,
+              sourceConfig: config,
+              hash: "model-providers-agent-credentials",
+              issues: [],
+              raw: JSON.stringify(config),
+              valid: true,
+            },
+            "config.patch": { ok: true },
+            "models.list": {
+              models: [{ id: "gpt-5.5", name: "GPT-5.5", provider: "openai", available: true }],
+            },
+            "models.authStatus": {
+              ts: now,
+              providerCapabilities: [
+                { provider: "openai", apiKeySupported: true, quickApiKeySetup: true },
+                { provider: "google", apiKeySupported: true, quickApiKeySetup: true },
+              ],
+              providers: [
+                {
+                  provider: "openai",
+                  displayName: "OpenAI",
+                  status: "static",
+                  profiles: [],
+                  apiKey: { source: "config" },
+                },
+              ],
+            },
+          },
+        });
+
+        await page.goto(`${suite.server.baseUrl}settings/model-providers`);
+        const openaiCard = page.locator('[data-provider-id="openai"]');
+        await expect.poll(async () => openaiCard.textContent()).toContain("Credentials for Main");
+        await openaiCard.getByRole("button", { name: "Replace key" }).click();
+        if (recordVisuals) {
+          await mkdir(artifactDir, { recursive: true });
+          await page.screenshot({
+            path: path.join(artifactDir, "provider-credential-scope-before.png"),
+            fullPage: true,
+          });
+        }
+        await openaiCard.getByLabel("API key").fill("synthetic-main-provider-key");
+
+        const agentPicker = page.locator(".agent-scope-control openclaw-agent-select");
+        await agentPicker.locator(".agent-select__trigger").click();
+        await agentPicker.locator('wa-dropdown-item[aria-label="Writer"]').click();
+        await expect.poll(async () => openaiCard.textContent()).toContain("Credentials for Writer");
+        await expect.poll(async () => openaiCard.locator('input[type="password"]').count()).toBe(0);
+        expect(await gateway.getRequests("config.patch")).toHaveLength(0);
+        if (recordVisuals) {
+          await page.screenshot({
+            path: path.join(artifactDir, "provider-credential-scope-inline-cleared.png"),
+            fullPage: true,
+          });
+        }
+
+        const addSection = page.locator(".settings-section", {
+          has: page.getByRole("heading", { name: "Add provider" }),
+        });
+        await addSection.getByRole("button", { name: "Add provider", exact: true }).click();
+        await addSection.getByLabel("Provider").selectOption("google");
+        await addSection.getByLabel("API key").fill("synthetic-writer-provider-key");
+        await agentPicker.locator(".agent-select__trigger").click();
+        await agentPicker.locator('wa-dropdown-item[aria-label="Main"]').click();
+        await expect.poll(async () => openaiCard.textContent()).toContain("Credentials for Main");
+        await expect.poll(async () => page.locator(".model-providers__add-form").count()).toBe(0);
+        await expect.poll(async () => openaiCard.locator('input[type="password"]').count()).toBe(0);
+        expect(await gateway.getRequests("config.patch")).toHaveLength(0);
+        if (recordVisuals) {
+          await page.screenshot({
+            path: path.join(artifactDir, "provider-credential-scope-after.png"),
+            fullPage: true,
           });
         }
       },
