+++ ui/src/pages/model-providers/model-providers-page.test-support.ts
@@ -1,0 +1,236 @@
+import { vi } from "vitest";
+import type { GatewayBrowserClient } from "../../api/gateway.ts";
+import type { ModelsProbeResult } from "../../api/types.ts";
+import type { ApplicationContext, ApplicationGatewaySnapshot } from "../../app/context.ts";
+import type { DefaultModelSelection, ModelProviderLogoutTarget } from "./data.ts";
+import type { ModelProvidersData } from "./load.ts";
+import type { ModelProvidersRouteData } from "./route.ts";
+import "./model-providers-page.ts";
+
+export type ModelProvidersPageTestElement = HTMLElement & {
+  context: ApplicationContext;
+  updateComplete: Promise<boolean>;
+  busy: Record<string, boolean>;
+  data: ModelProvidersData | null;
+  addProvider: () => Promise<void>;
+  addProviderId: string;
+  addProviderKey: string;
+  addProviderOpen: boolean;
+  defaultsDraft: DefaultModelSelection | null;
+  keyDraft: string;
+  keyEditorProvider: string | null;
+  logout: (cardId: string, targets: ModelProviderLogoutTarget[]) => Promise<void>;
+  messages: Record<string, { kind: "success" | "error"; text: string; warning?: string }>;
+  pendingLogoutProvider: string | null;
+  probe: (cardId: string, providers: string[]) => Promise<void>;
+  probeResults: Record<string, ModelsProbeResult>;
+  refresh: (opts: { force: boolean }) => Promise<void>;
+  routeData: ModelProvidersRouteData | undefined;
+  requestUpdate: () => void;
+  saveDefaultModels: () => Promise<void>;
+  saveKey: (provider: string, configKey: string) => Promise<void>;
+  selectedAgentId: string;
+};
+
+export type AgentSelectElement = HTMLElement & {
+  onSelect: (value: string) => void;
+};
+
+export function deferred<T>() {
+  let resolve!: (value: T) => void;
+  const promise = new Promise<T>((resolvePromise) => {
+    resolve = resolvePromise;
+  });
+  return { promise, resolve };
+}
+
+export function createHarness(initialScopeId: string) {
+  let pendingAuthStatus: Promise<void> | null = null;
+  let releaseAuthStatus: (() => void) | null = null;
+  const deferNextAuthStatus = () => {
+    pendingAuthStatus = new Promise<void>((resolve) => {
+      releaseAuthStatus = resolve;
+    });
+    return () => releaseAuthStatus?.();
+  };
+  let usageStatus: unknown = { updatedAt: 1, providers: [] };
+  let usageStatusRejects = false;
+  const request = vi.fn(async (method: string): Promise<unknown> => {
+    switch (method) {
+      case "models.authStatus": {
+        if (pendingAuthStatus) {
+          const gate = pendingAuthStatus;
+          pendingAuthStatus = null;
+          await gate;
+        }
+        return {
+          ts: 1,
+          providers: [],
+          providerCapabilities: [
+            { provider: "anthropic", apiKeySupported: true, quickApiKeySetup: true },
+          ],
+        };
+      }
+      case "models.list":
+        return { models: [] };
+      case "config.get":
+        return { config: {}, hash: "hash" };
+      case "usage.status":
+        if (usageStatusRejects) {
+          throw new Error("usage.status unavailable");
+        }
+        return usageStatus;
+      case "sessions.usage":
+        return { aggregates: { byProvider: [] } };
+      default:
+        return {};
+    }
+  });
+  const snapshot: ApplicationGatewaySnapshot = {
+    client: { request } as unknown as GatewayBrowserClient,
+    phase: "connected",
+    offlineStable: false,
+    canvasPluginSurfaceUrl: null,
+    hello: null,
+    assistantAgentId: "main",
+    sessionKey: "main",
+    lastError: null,
+    lastErrorCode: null,
+  };
+  const gatewaySource = publishableGateway(snapshot);
+  let selectionListener: (() => void) | undefined;
+  const agentSelection = {
+    state: {
+      selectedId: initialScopeId as string | null,
+      scopeId: initialScopeId as string | null,
+    },
+    set: vi.fn(),
+    setScope: vi.fn(),
+    subscribe(listener: () => void) {
+      selectionListener = listener;
+      return () => {
+        selectionListener = undefined;
+      };
+    },
+  };
+  const subscribe = () => () => undefined;
+  const runtimeConfig = {
+    state: {
+      connected: true,
+      configSnapshot: { config: {} },
+      configForm: {
+        agents: { defaults: { thinkingDefault: "low", fastModeDefault: "auto" } },
+      },
+      configLoading: false,
+      configSaving: false,
+      configApplying: false,
+      configNeedsApply: false,
+      configFormMode: "form",
+      configFormDirty: false,
+      configAutoSaveStatus: "idle",
+      lastError: null as string | null,
+    },
+    ensureLoaded: vi.fn(async (): Promise<void> => undefined),
+    patch: vi.fn(async () => true),
+    patchForm: vi.fn(),
+    removeFormValue: vi.fn(),
+    refresh: vi.fn(async () => undefined),
+    save: vi.fn(async () => true),
+    apply: vi.fn(async () => true),
+    discardDraft: vi.fn(async () => undefined),
+    subscribe,
+  };
+  const context = {
+    gateway: gatewaySource.gateway,
+    agents: {
+      state: {
+        agentsList: {
+          defaultId: "main",
+          mainKey: "main",
+          scope: "project",
+          agents: [
+            { id: "main", name: "Main" },
+            { id: "writer", name: "Writer" },
+          ],
+        },
+        agentsLoading: false,
+        agentsError: null as string | null,
+      },
+      ensureList: vi.fn(),
+      refreshList: vi.fn(),
+      subscribe,
+    },
+    agentSelection,
+    runtimeConfig,
+    overlays: {
+      snapshot: { updateRunning: false, updateReconciliationPending: false },
+      subscribe,
+    },
+    navigate: vi.fn(),
+  } as unknown as ApplicationContext;
+  return {
+    agentSelection,
+    context,
+    deferNextAuthStatus,
+    notifySelection: () => selectionListener?.(),
+    request,
+    runtimeConfig,
+    snapshot,
+    publishPhase: (phase: ApplicationGatewaySnapshot["phase"]) => {
+      snapshot.phase = phase;
+      gatewaySource.publish({ ...snapshot });
+    },
+    setUsageStatus: (value: unknown) => {
+      usageStatus = value;
+    },
+    failUsageStatus: () => {
+      usageStatusRejects = true;
+    },
+  };
+}
+
+export function publishableGateway(initial: ApplicationGatewaySnapshot) {
+  let current = initial;
+  const listeners = new Set<(value: ApplicationGatewaySnapshot) => void>();
+  return {
+    gateway: {
+      get snapshot() {
+        return current;
+      },
+      subscribe(listener: (value: ApplicationGatewaySnapshot) => void) {
+        listeners.add(listener);
+        return () => listeners.delete(listener);
+      },
+    },
+    publish(next: ApplicationGatewaySnapshot) {
+      current = next;
+      for (const listener of listeners) {
+        listener(next);
+      }
+    },
+  };
+}
+
+export function requestCount(request: ReturnType<typeof vi.fn>, method: string): number {
+  return request.mock.calls.filter(([candidate]) => candidate === method).length;
+}
+
+export async function advanceUsageRetries(): Promise<void> {
+  for (let attempt = 0; attempt < 3; attempt += 1) {
+    await vi.advanceTimersByTimeAsync(5_000);
+  }
+}
+
+export function focusDocument(): void {
+  vi.spyOn(document, "hasFocus").mockReturnValue(true);
+  vi.spyOn(document, "visibilityState", "get").mockReturnValue("visible");
+}
+
+export function appendPage(context: ApplicationContext) {
+  const page = document.createElement(
+    "openclaw-model-providers-page",
+  ) as ModelProvidersPageTestElement;
+  page.context = context;
+  document.body.append(page);
+  return page;
+}
